refactor: split command management into modules
The enormous `main` module was becoming very unwieldy and was difficult for AI to digest.

This splits the command handling into different modules.

A few things were removed in this refactor:

* Duplicate `DeployOptions` struct
* The `faucet` subcommand
* The `logstash` subcommand

A few functions have a Clippy suppression for too many arguments: it's not worth breaking these
arguments into an options struct.
